Refactoring: Renaming awaiting-amend to amendment-explanatory-page
