+++ app/middleware/submissionMiddleware.js
@@ -5,7 +5,7 @@ const { isToggleOnAwaitingAmend } = require('app/core/utils/checkToggle');
 const APPLICATION_SUBMITTED_PATH = '/application-submitted';
 const DONE_AND_SUBMITTED = '/done-and-submitted';
 const APPLICATION_MULTIPLE_REJECTED_CASES_PATH = '/contact-divorce-team';
-const AWAITING_AMEND_CASE = '/awaiting-amend-case';
+const AMENDMENT_EXPLANATORY_PAGE = '/amendment-explanatory-page';
 
 const handleCcdCase = (req, res, next) => {
   const session = req.session;
@@ -40,7 +40,7 @@ const handleCcdCase = (req, res, next) => {
   case 'AwaitingAmendCase':
     if (isToggleOnAwaitingAmend(session)) {
       logger.infoWithReq(req, 'awaiting_amend_case', 'Awaiting amend case');
-      return res.redirect(AWAITING_AMEND_CASE);
+      return res.redirect(AMENDMENT_EXPLANATORY_PAGE);
     }
     logger.infoWithReq(req, 'case_done_and_submitted', 'Default case state - redirecting to done and submitted');
     return res.redirect(DONE_AND_SUBMITTED);
