+++ app/steps/amendment-explanatory-page/schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "http://json-schema.org/draft-04/schema#",
-  "id": "awaiting-amend",
+  "id": "amendment-explanatory-page",
   "type": "object",
   "properties": {}
 }
