+++ app/steps/amendment-explanatory-page/template.html
@@ -1,7 +1,7 @@
 {% from "warning-text/macro.njk" import govukWarningText %}
 {% from "phase-banner/macro.njk" import govukPhaseBanner %}
 {% from "hmcts-progress-bar/macro.njk" import hmctsProgressBar %}
-{% from "common/components/awaitingAmendsComponent.html" import documentList, nextStep %}
+{% from "common/components/amendmentExplanatoryPageComponents.html" import documentList, nextStep %}
 
 {% extends "layout.html" %}
